Fix issues in Renderer and MainLoop
Missing super().__init__() call in MainLoop.

Renderer width must be set as external property and the spacer
calculated when this property is set. We need it to be able to create
default renderer in the App class.

Renderer had a bad _redraw set to False after screen was refresh and
showed, so when screen closed itself the next screen wasn't properly
drew.

Renderer's _process_input() method now getting the active_screen as
parameter.