Fix modal screen by removing execute_new_loop
Modal screen here is not modal from the point of interaction with other
screens. We always operate with just a one screen. Meaning of modal here
is about drawing this screen immediately. For this we don't need a new
event loop we just need to start a new redraw.

The method `execute_new_loop` was here just for modal window, so
removing everything which complicated code because of this feature.