Add temporal pylint-disable for with Lock
Lock from threading module is context manager which means it can be used
with `with` statement. This is false positive for pylint which is fixed
in python3-astroid package in version 1.5.3.

The python3-astroid 1.5.3. package is already in Fedora. However, it
contains old version of astroid thanks to packaging error.

When this will be solved we can revert this PR.