Do not run an old event queue from modal screen
When we end the event loop politely (self._run_loop) the modal screen
process signals in its own queue (which is correct) and then it's
going back through recursion calls of `processed_signals()` methods
however in this point there is an old event loop which will be
processed on the old screen. This may lead to unexpected behavior.

This issue starts to be a big problem for follow up commit which will
change how the modal screen is ended.