List containers take max width if not specified
ListRowContainer and ListColumnContainer could have unspecified width.
They will take a maximal possible width and distribute this to columns.

List containers will raise an exception if it is not possible to render
the container content. Or give a user hint to turn off numbering if it
can be rendered without the numbering.