Don't block loop when waiting on user input
The new get_user_input_async() method added to be able to ask for the
user input without blocking the main loop or running process_events().

This method is called as default for getting input from user for the
screens input() method.

As a bonus side effect this will reduce recursion calls of the
process_events() method.

Tests were modified to reflect these changes.