Move and rename force_quit loop
Force quit application is now a way how to kill application when
uncaught exception raised and it's not processed by a user. It was
moved to the AbstractEventLoop class and used by other loops. The
force kill application is now consistent for all the loops.

Rename the method to kill_app_with_traceback() and it calls sys.exit.

Print screen stack after exception.

Change tests to work with this solution and add glib variants.