Allow to disable concurrency check for all UIScreen inputs
Right now it is possible to disable concurrency check for InputHandler. However,
this class is created on demand in the InputManager used by the UIScreen which
means that a users are not able to disable concurrency_check without
re-implementation of the InputManager and InputHandler which is really cumbersome.

To solve this I'm adding a new property to InputManager and enable obtaining the
InputManager instance from the UIScreen.

The concurrency check shouldn't be disabled if you don't have a strong reason to
do it. It was implemented because without the check it's really hard to debug
when multiple screens are asking for the input at once. The main reason to disable it
is for error reporting after the application has crashed.

Related: rhbz#1807491